[#27]fix: globalException 위치 변경
